backend/requirements.txt
fastapi
uvicorn
pydantic
langgraph
langchain-google-genai
google-cloud-secret-manager
python-dotenv
httpx
